finished with the functionality of using saved templates rides

+++ controllers/callback-query.js
@@ -6,10 +6,23 @@ import { handlerMainMenuWeather } from '../keyboards/location_weather/handler-ma
 
 export async function callbackQuery(ctx) {
 	try {
+		const messageForDel = ctx.session.messageDel;
+		//если не был запущен rideon, то возврат
+		if (!messageForDel)
+			return await ctx.reply('Что то пошло не так, попробуйте заново запустить /rideon');
+
+		messageForDel.forEach(messageId => {
+			ctx.deleteMessage(messageId);
+		});
+		//обнуление массива сообщений для удаления
+		ctx.session.messageDel = [];
 		const cbqData = ctx.update.callback_query.data;
-		console.log(cbqData); //❗❗❗ for dev
-		// удаление меню инлайн клавиатуры после нажатия любой кнопки
-		await ctx.deleteMessage(ctx.update.callback_query.message.message_id);
+		// console.log(cbqData); //❗❗❗ for dev
+		// удаление меню инлайн клавиатуры после нажатия любой кнопки, исключение при выборе шаблонов поста
+		if (!(cbqData.includes('postId_') || cbqData.includes('_pattern_back'))) {
+			await ctx.deleteMessage(ctx.update.callback_query.message.message_id);
+		}
+
 		//обработчики меню создания объявления о заезде(поста)
 		await handlerMainMenu(ctx, cbqData);
 		await handlerSubMenu(ctx, cbqData);
